% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.Beta.LogLik}
\alias{PH.Beta.LogLik}
\title{Profiled log-likelihood function in Cox proportional hazards model}
\usage{
PH.Beta.LogLik(bet, yobs, delta, X)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates}
}
\description{
Calculate profiled log-likelihood function in Cox proportional hazards model.
}
