% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.fit}
\alias{PH.fit}
\title{Cox proportional hazards model based on the profiled likelihood estimation procedure}
\usage{
PH.fit(
  yobs,
  delta,
  X,
  Var = TRUE,
  Var.Robust = FALSE,
  eps = 1e-06,
  maxit = 50000
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates}

\item{Var}{a logical value. The default setup is \code{TRUE}, indicating that the standard errors of the estimated regression coefficients will be calculated.}

\item{Var.Robust}{a logical value. The default setup is \code{TRUE}, indicating that we assume that the model is correctly specified, so that there is no need to use the robust estimator of standard errors.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{res}.
}
\description{
Fit Cox proportional hazards model based on the profiled likelihood estimation procedure.
}
