% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.Influence.Lam}
\alias{PH.Influence.Lam}
\title{Influence functions for the estimator of the nonparametric part in the Cox proportional hazards model}
\usage{
PH.Influence.Lam(tm, yobs, delta, X, bet = NULL)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{bet}{unknown parameters corresponding to the model.}
}
\description{
Calculate the influence functions for the estimator of the nonparametric part in the Cox proportional hazards model.
}
