% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.Influence}
\alias{PH.Influence}
\title{Influence functions for the estimator of regression coefficients in the Cox proportional hazards model}
\usage{
PH.Influence(yobs, delta, X, bet = NULL)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{bet}{unknown parameters corresponding to the model.}
}
\description{
Calculate the influence functions for the estimator of regression coefficients in the Cox proportional hazards model.
}
