% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.Lam}
\alias{PH.Lam}
\title{Nonparametric component in Cox proportional hazards model}
\usage{
PH.Lam(tm, yobs, delta, X, bet = NULL, type = "right")
}
\arguments{
\item{tm}{The time points that the nonparametric baseline function will be estimated at.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{bet}{unknown parameters corresponding to the model.}

\item{type}{if \code{type="right"}, right-continuous results will be calculated, and if \code{type="left"}, left-continuous results will be calculated.}
}
\description{
Nonparametric component for the Cox proportional hazards model at specified time points.
}
