% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH_Online.R
\name{PH.Renew.Batch}
\alias{PH.Renew.Batch}
\title{Online estimation in Cox proportional hazards model}
\usage{
PH.Renew.Batch(
  yobs,
  delta,
  X,
  initial = TRUE,
  prev = NULL,
  maxit = 1000,
  eps = 1e-06
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{initial}{a logical value. The default specification is \code{TRUE}, indicating that the current data batch is the first batch and no historical summary statistics is available.}

\item{prev}{indicates the historical summary data. The default specification is \code{NULL} (with \code{initial=TRUE}). Otherwise, it is a list with the following five elements:
\code{bet} is the historical estimated result of regression coefficients;
\code{InfoM} is the historical estimated result of the information matrix;
\code{N} is the sample of historical raw data;}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{res}.
}
\description{
Fit online Cox proportional hazards model.
In other words, this function can analyze streaming survival data.
}
\examples{
## ---- generate the dataset (full) ---- ##
N <- 10000
sdata.full  <- sdata.PH(N=N)
yobs  <- sdata.full$sdata$yobs
delta <- sdata.full$sdata$delta
X     <- as.matrix(sdata.full$sdata$X)
rm(sdata.full)

## ---- fit the Cox proportional hazards model in an online manner ---- ##
# prepare basic elements
B <- 40
Nb <- N/B
batch <- ceiling((1:N) / Nb)
Res.Online <- array(0,dim=c(B,3,2),
                    dimnames=list(paste("batch",1:B,sep=""),paste("bet",1:3,sep=""),c("EST","SE")))
# the online procedure
for(b in 1:B){  # b <- 1
  cat("Batch",b,"...\n")

  # preparation: the batch idx / the previous elements
  idxb <- batch==b
  if(b==1){ prevb <- NULL }else{
    prevb <- c(
      fitb$fit
    )
  }

  # fit the current data batch (with current data the historical statistics)
  fitb <- PH.Renew.Batch(
    yobs=yobs[idxb],delta=delta[idxb],X=X[idxb,,drop=F],
    initial=(b==1),prev=prevb)
  Res.Online[b,,1] <- fitb$res[,1]
  Res.Online[b,,2] <- fitb$res[,2]

}
# present the fitted results
print(Res.Online)

}
