% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH_Online.R
\name{PH.Renew.Beta.LogLik}
\alias{PH.Renew.Beta.LogLik}
\title{Loss function for bet in Cox proportional hazards model (Online version)}
\usage{
PH.Renew.Beta.LogLik(
  bet,
  yobs,
  delta,
  X,
  bet.prev,
  InfoM.prev,
  ScoreAppro.prev,
  N.prev
)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{bet.prev}{the historical summary data of regression coefficients.}

\item{InfoM.prev}{the historical summary data of the information matrix (inverse variance-covariance matrix).}

\item{ScoreAppro.prev}{the historical summary data of a term derived from the online variable selection procedure (the recovering term).}

\item{N.prev}{the sample of historical raw data}
}
\description{
Loss function for the Cox proportional hazards model (Profiled online log-likelihood function in NPMLE).
Note that it highly depends on the function \code{\link{PH.Beta.LogLik}} in offline scenario.
}
