% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH_Online_Hetero.R
\name{PH.Renew.Hetero.ALasso.Batch.fit}
\alias{PH.Renew.Hetero.ALasso.Batch.fit}
\title{Online estimation in Cox proportional hazards model with dynamic covariate effects (without bootstrap rocedure)}
\usage{
PH.Renew.Hetero.ALasso.Batch.fit(
  yobs,
  delta,
  X,
  initial = TRUE,
  prev = NULL,
  tunpara.num = 50,
  tunpara.min.ratio = 1e-04,
  tunpara = NULL,
  evatype = c("AIC", "BIC")[2],
  learnrate = c(0.4),
  maxit = 1000,
  eps = 1e-06,
  threshold = 5e-05,
  trace = TRUE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{initial}{a logical value. The default specification is \code{TRUE}, indicating that the current data batch is the first batch and no historical summary statistics is available.}

\item{prev}{indicates the historical summary data. The default specification is \code{NULL} (with \code{initial=TRUE}). Otherwise, it is a list with the following five elements:
\code{bet} is the historical estimated result of regression coefficients;
\code{bet.batch} is the historical estimated result of regression coefficients using only the previous dataset;
\code{InfoM} is the historical estimated result of the information matrix;
\code{N} is the sample of historical raw data.}

\item{tunpara.num}{the number of tuning parameter values that will be fitted at.}

\item{tunpara.min.ratio}{Smallest value for tuning parameter, as a fraction of tunpara.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero).}

\item{tunpara}{A user supplied \code{tunpara} sequence. Typical usage is to have the program compute its own \code{tunpara} sequence based on \code{tunpara.num} and \code{tunpara.min.ratio}. Supplying a value of \code{tunpara} overrides this.}

\item{evatype}{the information criterion that will be used in the selection of optimal tuning parameter.}

\item{learnrate}{the learning rate in solving the optimization problem.}

\item{maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{maxit = 5e3}.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{threshold}{a threshold value. The deafult value is \code{1e-5}. If the absolute value of the each component of coefficients is smaller than \code{threshold}, it will be converted into zero.}

\item{trace}{a logical value. The default specification is \code{TRUE}, indicating that several useful information about the current process of fitting will be displayed.}
}
\value{
The fitted results are returned (a list). The estimates of regression coefficients is \code{resbet}.
}
\description{
Conduct Online estimation in Cox proportional hazards model with dynamic covariate effects (without bootstrap rocedure).
}
