% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH_Online_Hetero.R
\name{PH.Renew.Hetero.betTau.Loglik}
\alias{PH.Renew.Hetero.betTau.Loglik}
\title{Loss function in Cox proportional hazards model (Online, Heterogeneity)}
\usage{
PH.Renew.Hetero.betTau.Loglik(
  bet,
  tau,
  yobs,
  delta,
  X,
  bet.prev,
  InfoM.prev,
  N.prev
)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{tau}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates}

\item{bet.prev}{the historical summary data of regression coefficients.}

\item{InfoM.prev}{the historical summary data of the information matrix (inverse variance-covariance matrix).}

\item{N.prev}{the sample of historical raw data}
}
\description{
Calculate the loss function in Cox proportional hazards model (Online, Heterogeneity).
}
