% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.Score.Individual}
\alias{PH.Score.Individual}
\title{Individual level score vector in Cox proportional hazards model}
\usage{
PH.Score.Individual(bet, yobs, delta, X, Iscov = TRUE)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{Iscov}{whether convert the individual level form into a covariance form.}
}
\description{
Calculate the individual level score vector in Cox proportional hazards model.
}
