% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PH.R
\name{PH.ScoreInfoM}
\alias{PH.ScoreInfoM}
\title{Score vector and information matrix in Cox proportional hazards model}
\usage{
PH.ScoreInfoM(bet, yobs, delta, X, IsScore = FALSE, IsInfoM = TRUE)
}
\arguments{
\item{bet}{unknown parameters corresponding to the model.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates.}

\item{IsScore}{whether the score vector is calculated or not.}

\item{IsInfoM}{whether the information matrix is calculated or not.}
}
\description{
Calculate score vector and information matrix in Cox proportional hazards model.
}
